[WIP] Change error type for Eval::eval and Value::funcall
This commit does not compile.

Begin refactoring the VM to return raiseable errors from functions that
evaluate expressions on the interpreter. This will allow exceptions to
be propagated through native code without an unwind, which is more
ergonomic and removes a large source of potential Undefined Behavior.

This PR alters `Eval::eval` and `Value::funcall` to return
`Box<dyn RubyException>`. This required changes to `artichoke-core` to
parameterize error types with an associated type in the trait
definitions.

This change obsoletes `Eval::unchecked_eval` and
`Value::unchecked_funcall`. These trait prototypes and their
implementations in `artichoke-backend` have been removed.

`Value::freeze` and `Value::respond_to` return the raiseable error type
to enable propagating interpreter exceptions.

This commit alters the `ExceptionHandler` trait to have this prototype:

    fn last_error(&self) -> Result<Option<Exception>, ArtichokeError>;

`Exception` is a newly raiseable type that implements `RubyException`
and is derived from an `Exception` `mrb_value`.

Backtraces are disabled for now.