Return Option<NonNull<sys::rclass>> from rclass APIs.
This commit breaks the dependency of `def` module on the Artichoke
interpreter. `def` is predominantly glue between Rust and FFI and only
needs a reference to the `mrb_state`.

`class::Spec::rclass` and `module::Spec::rclass` now take an
`mrb_state`.

`module::Spec::new` takes an `Artichoke` and eagerly generates the
`Symbol` ID for the class name, which removes the `Cell` member and
enables `module::Spec::rclass` to take only an `mrb_state`.