Push nil-value coercion to the edge of Array API
Change the return type on ArrayType::get to return Option<Value> in
the success case. This makes the type more closely align with std
collections and may eliminate the need for an ArrayType to use the
converter APIs.

This pushes conversion of None valued options into a nil Value to
the public API of Array.

This is part of a push to push interpreter glue to out of core type
backends.