Implement default fill branches of Array#initialize constructor
When InlineBuffer was introduced, the branches in Array::initialize that
were previously implemented with `backend::repeated::value` and
`backend::fixed::hole` were replaced with a call to `unimplemented`.

This commit implements these cases and does so by unifying them into one
branch. This branch and the one above for block initializations are
modified to build a buffer of `sys::mrb_value` instead of `Value` to
reduce an allocation while constructing the `SmallVec`.