Rewrite random module to introduce a new state component
This commit renames `random::backend::Rand` trait to `RandType` to match
more closely with other modules. The API has been redefined to remove
fallibility and take mutable references to the interpreter when
required. This approach is part of gradually introducing mutability
requirements that was first rolled out in GH-467.

The previous implementation of the interpreter default PRNG attached a
`Random` instance to the `State`. This approach is not compatible with
an idiomatic borrowing setup, as found in GH-442.

This PR introduces a new `Prng` state component. The `Prng` binds
directly to a `Rand<SmallRng>` instead of a `Random` wrapper around a
`RandType` trait object.

`Rand` random backend has now implemented methods outside of `RandType`
that do not take an interpreter. Since the PRNG on the state binds
directly to a `Rand`, this breaks the circular dependency on the
interpreter.