Remove artichoke-array feature
The `artichoke-array` feature was added to enable a temporary rollout of
the Artichoke-native `Array` to the playground as the implementation
evolved and shook off all of the segfaults.

As of GH-382, these segfaults are fixed since the Array type is
implemented with a contiguous buffer.

This commit removes the legacy code, makes the Artichoke Array the
default, and removes the feature.

As of this commit, mruby is always built with `-DARTICHOKE`. This change
was required as I think about implementing more core types, like
`Symbol`.