Reset #[must_use] annotations
Clippy in Rust 1.40.0 was noisy with suggesting `#[must_use]`, requiring
them on trait impls. The result was a lot of noise in the code and toil
when writing code.

The lint has been reworked for Rust 1.41.0. This commit removed every
`#[must_use]` annotation and reconverged based on the new lint.

The result was to delete a significant number of annotations from the
source.