Remove macro-generated converter implementations
Remove the 1000+ Convert and TryConvert implementations. Remove the
`artichoke-all-converters` feature.

Manually implement all converters required by the current application.

Converters are specialized for each type and do not funnel through the
base `Value` converter for collections.

This change is in preparation for converter work falling out of GH-442
which will require converters that allocate on the interpreter heap as
`&mut self`.

This patch removes a bunch of code and speeds up compile times. (Shaves
5s or 10% off a clean build on my machine.)

Fixes GH-486.