Replace crates.io rust-onig with the artichoke fork
The playground has long used a fork of rust-onig that lives at

https://github.com/artichoke/rust-onig

The playground used this fork to pull in changes to the `onig_sys`
`build.rs` to add wasm support.

artichoke/rust-onig@952305049073f37ada5543536c4c58540ea3642b adds a
change to the `bindgen` dependency of `onig_sys` to pull in the same
version of `bindgen` that is used in `artichoke-backend`.

This change reduces the dependency count in `Cargo.lock` measurably.
With this commit, `Cargo.lock` contains only one version of `syn`,
`quote`, and `proc-macro2`. Artichoke no longer has `failure` or
backtrace` in its dependency tree.