Remove ArtichokeError from `def`, `class`, and `module` modules
Rather than return `ArtichokeError` in  `Result`s, instead return either
`ConstantNameError` or `NotDefinedError`.

`ConstantNameError` is a ZST that indicates the supplied constant (e.g.
a class, module, or method name) contains a NUL byte. In the future this
error type would indicate other parsing failures, e.g. a method name
starting with `;`.

`NotDefinedError` is an enum that indicates some expected part of the
class or constant heirarchy was not defined. For example, when defining
a class that inherits, if the super class does not exist, an error is
returned. In practice, this error type defends against initialization
order issues and FFI failures (returned NULL pointers).

`ConstantNameError` and `NotDefinedError` implement `std::error::Error`,
`RubyException`, and a full compliment of `From` conversions to allow
these errors to be used ergonomically with `?` operator in Artichoke
where `Exception` is expected, for example, when building types in
`extn`.

`ConstantNameError` raises as a `NameError` in Ruby. `NotDefinedError`
raises as a `ScriptError` in Ruby.

This commit changes the `InitializeResult` type alias in `extn::prelude`
to use `Exception` as the error type. Some call sites that explicitly
construct `Artichoke::New` were reqwritten to use the `NotDefinedError`
type.

This commit removes fallibility from `Method::define` since the error
variant was never constructed.