Reimplement artichoke-backend VFS with a HashMap
The virtual filesystem used to be implemented with the artichoke-vfs
crate, a fork of the filesystem crate. This fork added support for
custom metadata to the filesystem nodes. The metadata stored Rust
require hooks (Rust extensions analagous to C RubyGems in MRI) and
whether the source had been required before for short circuiring in
`Kernel#require`. The VFS itself was used to store UTF-8 Ruby source
file contents.

This commit replaces that implementation with a
`HashMap<PathBuf, Entry>`. `Entry` contains all of the required bit,
optional file contents, and optional Rust extension. Content is only
modifyable by the `LoadSources` APIs. File contents themselves can only
be replaced; they may not be modified in place.

The VFS tracks a current working directory, but does not expose any way
to modify it. CWD is set to `RUBY_LOAD_PATH`.

The new VFS implementation returns native `io::Result` from Rust `std`.

This commit wires the new VFS in through the `State` and interpreter
constructor function.