Remove ArtichokeError and BootError
This completes the migration to a unified error type in Artichoke that
began with GH-311. All uses of the `ArtichokeError` enum have been
migrated to specialized error types that implement `std::error::Error`
and `RubyException`. This enables them to be handled by standard Rust
error handling mechanisms as well as be raisable on an Artichoke
interpreter.

`BootError` was always a temporary stopgap to bridge the
`ArtichokeError` and `Exception` error types in frontend and `extn` code
as the rollout and migration was executed.

The following PRs contributed to this fix:

- GH-311
- GH-371
- GH-387
- GH-430
- GH-437
- GH-457
- GH-461
- GH-463
- GH-465
- GH-470
- GH-474
- GH-496
- GH-497
- GH-500
- GH-507
- GH-508
- GH-509
- GH-510
- GH-511
- GH-512