Remove rayon from artichoke-backend build
`rayon` is a heavy dependency and brings in a lot of crossbeam crates.
Depending on `rayon` directly in artichoke-backend's build.rs used to be
free because `cc` used rayon to implement the `parallel` feature for
parallel compilation. cc 1.0.42 moved away from `rayon` to `jobserver`.

This commit removes `rayon` and converts the `par_iter` loops into normal
for loops.

This code also cleans up the package generation code:

- renamed functions for path generation
- renamed erb invocation to "module render" step
- simplified source iterator chain
- pass `--disable-all` flag to MRI to disable all features instead of
  only `gems` and `did_you_mean`

Added documentation to birthdate build step that the code assumes there
is only one reachable git root.

This commit also removes the explicit `bstr` dependency in
`artichoke-frontend` that was unused.