Introduce Value::implicitly_convert_to_string
And migrate all explicit callers of `to_str` to use it except for one
problemetic case in `matchdata`.

Also introduce an `implicitly_convert_to_nilable_string` for use in the
`Regexp` code.

This refactors the existing `implicitly_convert_to_int` to return the
more specialized `TypeError`.

This commit removes several unnecessary uses of `format!` which removes
some panicking pathways. String concatenation and interpolation can be
done with a mutable `String` and `push_str`.