Add sys::mrb_sys_state_debug free function
This function replaces the DescribeState trait and the fmt::Debug
implementations on the various mrb_state pointer types.

DescribeState was introduced because it was not possible to implement
fmt::Debug on a *mut sys::mrb_state due to conflicts with blanket impls
in std.

The trait was awkward to use and provides no additional value over a function
that takes a *mut mrb_state directly.

This removes some panicking pathways from sys and makes the debug info generation
infallible and panic free.