Add GitHub actions workflows for building Artichoke
This commit adds two workflows: `ci` and `rustdoc`. The `rustdoc`
workflow is equivalent to the `deploy` job in the CircleCI config. The
`ci` job splits apart by source language with a dedicated job for
building and running tests of the rust sources.

The Rust build and test suite are run on a matrix of linux, windows, and
macos. The remaining jobs (linting and formatting for the various
languages) run on only linux.