Log stdout AND stderr on failed subprocesses in build.rs
build.rs does not care about the output of some external commands. Using
the `Command::status` API for these ensures that stdout and stderr are
inherited from the build.rs script, which means debug output is piped
through by default.

For the few cases where we care to consume the output of the subprocess,
write a summary of the stdout and stderr pipes using `eprintln!` and
panic with a simple error message, relying on verbose output from cargo
if the build script fails.