Switch to Rake for dev environment orchestration
This change moves away from Node.js and Yarn for orchestrating the
linter and formatter framework in Artichoke. Dev task runner moves from
node scripts and package.json inline scripts to Rake.

Rake is a Ruby task runner. Since Ruby is already a dependency of
Artichoke (and Artichoke is a Ruby), this is a more reasonable
dependency to take.

Node and prettier are still used for formatting, but only of text
and C sources, which are edited more rarely. Prettier is now invoked via
the command line instead of the JS API. This is why there are changes to
`.prettierignore` and `.editorconfig`.

This commit adds text, c, and js jobs to the CI workflow.