Add script for formatting C sources with clang-format
This script is a specialization of the previous `scripts/lint.js`.

`clang-format.js` formats C and header files in parallel and rewrites
them in place OR exits on formatting errors if given the `--check` flag.

This script outputs the files it processes like the prettier CLI.

This script will be called directly from Rakefile and CI.