Use constant definition APIs from artichoke-core in extn
`Float` and `Regexp` both define constants by string formatting `i64`s
and `f64`s  into a code string with `format!` and evaling it on the
interpreter. This is not ideal.

GH-526 added constant definition APIs to `artichoke-core` and
implemented them for `artichoke-backend`. This commit wires it up in
`extn`. The resulting code has more localized failures - errors are
reported on the granularity of defining a single constant. The constant
values are converted into proper `Value`s directly from their native
Rust form.

See also GH-27, the issue referenced by both of the TODOs (the TODO in
`float.rs` refers to the issue number from lopopolo/ferrocarril before
it was migrated to the artichoke/artichoke repo).