Remove panicking branch from exception::raise
Remove a `panic!` from `exception::raise` by delegating to an mruby
lookup of the `RuntimeError` exception class which is infallible.

Long term, The Artichoke state should store a lazily initialized
reference to the `Fatal` exception type, which we would use here.

This commit uses `hint::unreachable_unchecked` instead of the panicking
`unreachable!` because this function is guaranteed to unwind.