Replace IO APIs on State with an output strategy
Replace `State::puts`, `State::print`, `State::capture_output`, and
`State::get_and_clear_captured_output` with an output strategy.

The output strategy is a public field on the state. Changing the
strategy means assigning a new one to this property.

A strategy has methods for writing to stdout and stderr and two inherent
methods for `puts` and `print` that write to stdout and suppress errors.

Two implemented strategies are `Process`, which uses the process
`io::stdout` and `io::stderr`, and `Captured`, which has a growable
buffer for each of stdout and stderr.