Replace path converters with bstr functions from ByteSlice
The `bstr` crate that `artichoke-backend` already depends on implements
converters between `&[u8]` and `&OsStr` in each direction. This replaces
the (identical) implementations in the `ffi` module.

Docs: https://docs.rs/bstr/0.2.11/bstr/#file-paths-and-os-strings

This offloads non-core functionality to a high-quality crate, which is a
maintainability win. This also removes some conditional compilation,
which is another maintainability win.

The public functions for the conversions in FFI remain because they wrap
the `bstr` calls in `Results` with `RubyException` `Err` variants.

This commit removes unnecessary roundtrip conversions of
`fs::RUBY_LOAD_PATH` in `require`.