Prepare Regexp module for global variable APIs
When prototyping the `Regexp` module with APIs for getting and setting
global variables, I made these changes as well. They are worthy of their
own commit and were making the diff hard to reason about.

- Rename all parameters named `pattern` that are really `haystack`.
- `regexp::nth_match_group` takes a `NonZeroUsize`.
- Remove `panic!` for `0` case in `regexp::nth_match_group`.
- Add `regexp::clear_capture_globals` API.
- `active_regexp_globals` on the interpreter `State` is now an
  `Option<NoNonZeroUsize>`.

I tried to make `regexp::nth_match_group` return a `Cow<'static, str>`
since I am planning for constant names to be `&str`, but translating
this `Cow` to the `[u8]` `Cow` expected by `intern_symbol` was invasive
and the change is expected to be undone. Skipping for now.