Replace all usage of format! macro with panic-free alternatives
Replace all usage of the `format!` macro in non-test and non-build code
to use either:

- mutable `String` builder with `push` and `push_str` for concatenation
  use cases.
- `string::format_int_into` + `String` builder for interpolating integer
  values.
- `string::write_float_into` + `String` builder for interpolating float
  values.

This PR adds a dependency on `dtoa` in artichoke-backend. This dep is
the float equivalent of `itoa` that was added in GH-547.

This PR continues the work in GH-547.

This type of formatting is predominantly used when building `Exception`
messages, so the performance speed up will not be noticable.

Filed dtolnay/dtoa#18 to investigate adding a `dtoa::fmt` variant to
make the `string` API similar to `format_int_into`, which can take a
`String`.

This commit removes all uses of `String::from_utf8_lossy` in Artichoke.