Remove downcast dependency and all use of Any
This commit removes the `downcast` dependency from `artichoke-backend`
and replaces all instances of downcasting trait objects with dynamically
dispatched APIs.

One common use for downcasting in Artichoke was to get a concrete type
to feed into a `debug_struct` invocation in a `Debug` implementation.
This pattern can be used by adding a `fn as_debug(&self) -> &dyn fmt::Debug`
method to the trait definition, allowing concrete implementations of a
trait to forward their debug impls.

`ArrayType` used downcasting to specialize some operations for
`InlineBuffer` types. `ArrayType` is kept for compatibility and for the
medium term I expect `InlineBuffer` to remain the only implementation.
This PR reallocates a temporary `InlineBuffer` to make the
implementation work for all `ArrayType`s.

`RandType` implementations used downcasting to check internal state for
equality. This use was replaced with an enum that can represent internal
state for all `RandType` implementations.

There are no more imports of `std::any::Any` in this workspace.

This change prevents generics from leaking out of backend
implementations, like `SmallRng` in the `Rand` backend.