Disable with-dirs cargo feature in rustyline
Disable `rustyline` `with-dirs` feature since it is unused in
`artichoke-frontend`. This change removes some dependencies of
`rustyline` from the `Cargo.lock`.

The `with-dirs` feature of `rustyline` activates tab completion behavior
for paths that start with `~`. `airb` uses `rustyline` to input Ruby
source code, not paths, so this optional feature is not being used.