Embed packages generated by rubylib build step
`artichoke-backend` vendors a copy of Ruby 2.6.3. This vendored copy is
used in `build.rs` to extract Ruby source files for stdlib packages and
generate a Rust module that includes them in the `Artichoke`
interpreter.

These files do not change ~ever since we rarely update our Ruby fork. It
is easier to revendor individual sources as they are modified upstream
and remove this heavy build step.