Hide rustdoc index asset override in .github directory
`assets/index.html` is only used by the `rustdoc` workflow.

This file exists because long ago the artichoke/playground lived in this
repo and the playground was deployed as the root document for this
repo's GitHub Pages build.

`index.html` does an HTML meta redirect from the
https://artichoke.github.io/artichok/ to https://artichoke.run/. This
does not impact rustdocs since, as a virtual workspace, there is no
ruistdoc `index.html` in this position (it requires a nightly option
that Artichoke does not use).

This commit hides this bit of legacy configuration in the GitHub Actions
configuration and modifies the `rustdoc` workflow to use the new
location.