Use i64 for Fixnum backend on all targets
This PR simplifies Artichoke's numeric tower by using `i64` as the
Fixnum backend on all targets. This removes conditional compilation of
the `Int` type alias for `wasm32` targets and adds tests to assert that
`Int` is `i64`.

This change simplifies the implementation of integer converters by
allowing all targets to use lossless converters for `u32`. This change
is propagated to all `extn` callsites and partially undoes GH-596.

`build.rs` is modified to always pass `MRB_INT64` when compiling mruby
so Artichoke can safely assume all builds use `i64` `Fixnum`.

Because this change removes a source of platform-divergence, it helps
reduce the need for GH-413.

This PR cleans up the tests for the fixnum converter and annotates all
of the passthrough converters with `#[inline]`.