Experiment with using trampoline modules as a VM compat layer
This commit tries to formalize the idea of what the `trampoline` modules
in `extn` are useful for. I propose all core `extn` APIs accept native
types as parameters and return native types. `trampoline` modules are
responsible for boxing and unboxing values as the glue layer between the
core and the Ruby VM.

This commit tries this approach out on the `Integer` module.