Refactor RegexpType::scan to take and return native types
Change haystack parameter to a `&[u8]`.

Return a new enum `Scan` that can either tell the caller to return the
original haystack value, provide a list of captures, or a list of
matched subslices of the haystack.

This change simplifies scan implementations and does not add much
complexity to the `String#scan` Regexp callsite.

This commit fixes a bug where `$&` was not unset if there was no match.