Use Regexp trampoline as a VM compat layer
Followup to GH-607 by extending the native types to the `Regexp` API
instead of stopping at `RegexpType`. These changes are necessary for
refactoring `MatchData` to use modern patterns.

This PR uses the same tricks introduced in GH-610 to use the converter
infra to parse arguments out of a VM `Value`.

Related to GH-605.