Remove Rc wrapper from the State struct in Artichoke interpreter
A very early design choice (c39cc778900ff26d1ce8109f7235554e9e6c525a --
commit 70!) wrapped the Rust interpreter `State` in an `Rc<RefCell<_>>`.
This choice served the project well -- it mostly enabled ignoring pointer
lifetimes -- but now it is time to unwind this change.

I've been making a push to remove access to the `mrb_state` from the
interpreter and only expose access to high-level APIs. This concept has
existed for a while -- the traits that get implemented on the
`Artichoke` struct are those APIs. Things like `Intern`,
`DefineConstant`, and `Warn`.

One of the APIs I sought to add was a symbol table, which I intended to
use to reimplement the `Symbol` API in Rust and disable mruby's C
implementation.

The core challenge I ran into is that the `RefCell` wrapper around the
`State` prevented components of the `State` from being borrowed
independently (partial borrows on the `State`) which made composing
together the pieces of the state I was splitting out difficult or
impossible. GH-468 is a good example of this.

h2. Preparation

I attempted to unwind this mistake in GH-442, but this design choice had
infected the entire codebase. The `RefCell` and explicit borrows were
everywhere. GH-442 was a huge PR at over 2000 lines, but it was never in
a compiling state.

Instead, I started spinning out the good parts of GH-442 into their own
PRs, spawning over 50 of them as I went. These PRs included tactical
steps like using `NonNull` for parsing pointers in more places and
deleting dead code to limit how much to refactor. The remaining
strategic pieces were more intertwined.

h3. Everything takes `&mut Artichoke`

Part of the motivation to move away from `Rc<RefCell<_>>` was to get the
`Artichoke` struct to be covariant with respect to lifetimes. Before
this PR, the `Artichoke` struct contained a `*mut sys::mrb_state` which
does not require a mutable borrow to get unique access to the underlying
mruby interpreter. To achieve covariance requires storing a
`NonNull<sys::mrb_state>` instead -- getting at the underlying `*mut
sys::mrb_state` requires an `&mut self` receiver.

All of the core traits needed to be reworked, like in GH-480 and GH-622,
and all of `extn` needed to be migrated to pass mutable borrows of the
interpreter.

h3. Remove access to state internals in `extn`

The core and standard library was a mess of encapsulation breaks that
dug down into the interior of the `State`. There was work to remove
`sys` FFI calls from the core as well as pull accesses to the `State`,
like tracking `Regexp` globals, into these higher-level, trait defined
APIs.

Work in this area includes PRs like GH-469 and GH-566.

h2. Removing the `Rc<RefCell<_>>` contagion

This PR rips off the bandage and excises the `Rc<RefCell<_>>` from
Artichoke.

The approach was to redefine the `Artichoke` struct to look like:

```rust
struct Artichoke {
  pub mrb: NonNull<sys::mrb_state>,
  pub state: Option<Box<State>>,
}
```

The `State` is wrapped in an option to enable moving the `State` in and
out of the `Artichoke` wrapper as the stack enters and exits from C FFI.
When Rust code calls an mruby C function, the state is taken out of the
shell, turned into a raw pointer, and stored in the mruby userdata
pointer. If the mruby VM delegates to Rust code, pull the state out of
the userdata, convert it back to a `Box` and shove it into a new
`Artichoke` struct on this side of the FFI boundary. On exit, do the
whole thing in reverse. (See `Artichoke::with_ffi_boundary` and
`Guard::new`).

The `State` is now more difficult to use directly since it is wrapped in
an option. The introduction of the higher level API traits addresses
this. The only code that needs to access the `state` property directly
are those that implement the VM.

h3. New primitives

This PR contains a new `TypeRegistry` container in
`artichoke_backend::state` that backs new APIs for defining and
retrieving classes and modules. This change could not take place in the
preparation phase because mruby requires the retrieval APIs to return
references which the `RefCell` prevented.

Every interpreter API that was not already fallible, like the `Regexp`
global state APIs introduced in GH-658, has become fallible to support
the fatal error condition of the `State` being moved out in Rust code.

This PR makes much more consistent use of the `log` facade for emitting
`trace`, `debug`, and `error` logs related to the FFI movement process.

Value of `Class` and `Module` APIs as well as the new instance of class
API have been hoisted out of the `Spec`, which were always intended to
be low-level constructs, to the registry APIs.

Most converters have become the "mut" variety due to the state
covariance changes.

Class and module registry APIs significantly affected the `init`
functions in each extn module. The introduction of `Guard` touched every
mruby extern "C" trampoline.

h2. :tada:

Fixes GH-442.
Fixes GH-661.
Fixes GH-468.