Take mutable reference to an interpreter in RustBackedValue
To prepare for closing out GH-442, the data converters need access to a
raw interpreter, which will require exclusive access.

This PR takes care of the last stragglers in the extn trampolines that
did not take `&mut Artichoke`.

This PR introduces `TryConvertMut` implementations from `Value` to a
Rust type. This introduces a split in APIs that perform conversions
automatically based on a type parameter. This PR introduces
`Value::try_into_mut` but does not introduce the corresponding
`Value::funcall_mut`.