Convert filename argument in LoadSources from &[u8] to AsRef<Path>
In practice these APIs are called when a `Path` is already in context
which requires marshalling to bytes only to undo that by marshalling
back to `Path`.

Converting the filename argument to be a `Path` (and renaming it to
`path`) documents the intent that this argument will load from the
filesystem. It also pushes conversions to callers _only if they are
necessary_.

This PR is required to address GH-442, which will be lifting direct
access to the VFS out of `Kernel#require` into APIs in the `LoadSources`
trait.