Fix a use-after-free caused by GC cycles with the State moved out
Some Artichoke APIs allocate on the underlying mruby interpreter. When
an allocation happens, mruby may decide to run a garbage collection.
During a garbage collection, all `Array` values are marked. Since
`Array` is implemented in Rust, the `Array` GC mark routine is
implemented in Rust and expects to parse a well-formed interpreter.

Previously, the `State` was not moved back into the `mrb_state` userdata
pointer before these allocations. This means that the mark for `Array`
values would do nothing, all of the values in the `Array` would be freed
by the GC. Any subsequent uses of these values could segfault the
interpreter because of a use-after-free bug.

This commit adds a new and unsafe API to the interpreter called
`with_ffi_boundary`. This function takes a `FnOnce(*mut mrb_state) -> T`.
Before calling the closure, the API moves the `State` out of the
`Artichoke` and into the `mrb_state`. Inside the closure, only access to
the `*mut mrb_state` is allowed.

There are 2 instances where this approach cannot work due to the C APIs
requiring a borrow on the `State` to access a `mrb_data_type` from a
`class::Spec`. In these places -- `ClassRegistry::new_instance` and
`RustBackedValue::try_into_ruby` -- I'm using a hammer and temporarily
disabling the GC.