Ensure class and module specs are alive when running mrb_close
The sometimes segfault on Windows is caused by state.close() consuming the
State before running the final gc in mrb_close. The gc relies on the class
specs free function to free Rust-defined data types. This data is not live
after the state is consumed.

The fix is to destructure the state, pulling apart the bits that need explicit
closes and those that must be kept alive. The class and module registries are
explicitly dropped last.