Print exception backtraces for top-level errors in spec-runner
If an exception in the spec-runner happens outside of `MSpec`, it is
currently swallowed. This makes debugging difficult as we modify the
core VM.

This change catches errors returned from the interpreter and formats
them to stderr with the backtrace formatting routines in Artichoke
crate.

This change brings in a dependency in spec-runner on `termcolor`.