Update artichoke-core to be no_std + alloc compatible
`artichoke-core` defines the traits that comprise an interpreter. There
is little implementation. It is feasible to make the crate `no_std`. To
make this transition, the following changes were made:

- Add `extern crate alloc` to `lib.rs`.
- Add an optional, enabled by default, `std` feature.
- `use alloc::vec::Vec` where `Vec` was used since there is no longer
  any implicit `std` prelude.
- `use alloc::borrow::Cow` instead of `use std::borrow::Cow`.
- Remove `std::error::Error` bounds on all `Error` associated types.
- Update doc comments to refer to `core` instead of `std` where
  possible, with external links otherwise.
- Gate `Eval::eval_os_str` behind the `std` feature and note the feature
  requirement in the docs.
- Gate `Eval::eval_file` behind the `std` feature and note the feature
  reqirement in the docs.
- Use `str` instead of `Path` for path arguments in `LoadSources` when
  compiling for `no_std`.

This commit moves the `where Self: Sized` bound on `Value` from the
trait definition to the `self` converter methods.

This commit significantly improves the quality of the crate-level
documentation and README.