Reorganize preludes to avoid clippy glob import warnings
artichoke-core was exposed as a module in artichoke-backend's prelude
and was expected to be glob imported. Clippy only allows glob imports
on modules where the leaf name in the path is 'prelude'.

This commit includes all traits from artichoke_core::prelude in
artichoke_backend::prelude and artichoke::prelude.

This change touched the few doc tests that exist and updated the enclosing
functions from 'main' to 'example' and calls them to ensure the code is
executed.