Replace unsafe blocks in SmallArray::set_slice with safe alternatives
Replace the `unsafe` blocks that copy raw memory within the `SmallVec`
to do a multiple element remove with a call to `SmallVec::drain` which
is a safe API that does the same thing.

This commit also simplifies the range construction to pass to `drain`
and moves the `0` length slice case to the first case in the match to
minimize the possibility of bugs in other branches from impacting the
single drain case.