Improve docs and performance of AllSymbols iterator
Add additional documentation to `AllSymbols`, the `InternerAllSymbols`
extension trait, and `From` converters.

Implement more functions in the `Iterator` implementation to delegate to
optimized versions in `Range`. Manually implement `<AllSymbols as Iterator>::count
to work around a limitation in `std`.

Add tests for `AllSymbols` with `SYMBOL_RANGE_START` starting offsets.
And document the maximum number of elements that can be yielded from an
`AllSymbols` iterator.