Add Symbol::write_inspect_into
This API is in addition to `Symbol::format_inspect_into` and differs by
taking a `std::io::Write` instead of a `core::fmt::Write`. This method
UTF-8 encodes each yielded `char` from the `Inspect` before writing the
contents to the writer.

This API requires both the `artichoke` and `std` features to be active.