Fix handling of invalid UTF-8 and Unicode replacement character codepoints
The CharIndices iterator in bstr overloads the meaning of � when iterating.
This requires testing the byte span to determine if the replacement character
is legitimage or a substitute char.

This uncovered a bug where otherwise valid idents with invalid UTF-8 bytes
would be considered valid identifiers.

This commit adds tests for these edge cases.