Rename any exception error type in artichoke-backend
Rename `artichoke_backend::Exception` to `artichoke_backend::Error`.

This removes a name conflict with the base `Exception` class that is the
root of the Ruby error hierarchy. This exception type is explicitly not
exported in the artichoke-backend `prelude`.

This change aligns `artichoke_backend::Error` with the name of
`anyhow`'s error type. These two serve a similar purpose.

The motivation for this change is to make it easier to split all of the
exception types defined in `extn::core::exception` to a spinoso crate.

Most of this change is mechanical find and replace. There are some
places where additional edits were made:

- converters were updated to use `UnboxRubyError::new(...).into()` for
  constructing errors. This makes the code more update-proof.
- `artichoke-backend/src/exception.rs` renamed to `error.rs`.
- All `mruby.rs` in `extn` updated to call `error::raise`.
- `CaughtException` moved to `exception_handler.rs` since this is the
  only place it is used.
- Add a `Builder` for `CaughtException` to have a more ergonomic API for
  constructing the exception.
- Update `# Error` docs for `exception_handler::last_error`.
- `extn::core::env` was updated to use `EnvArgumentError::from(...).into()`
  for constructing errors. This makes the code more update-proof.
- `extn::core::math` was updated to use `NotImplementedError::from(...).into()`
  for constructing errors. This makes the code more update-proof.
- `gc::arena::IndexError` renamed to `ArenaSavepointError`.
- Ruby `Exception` type exported via `artichoke_backend::prelude`.
- Do not export `artichoke_backend::interpreter` or
  `interpreter_with_config` via the prelude.

NOTE: `ENV` implementation and backends is a good candidate to turn into
a spinoso crate.